feat(react-native-sdk): review remarks 2
