+++ react-native-sdk/android/src/main/java/org/jitsi/meet/sdk/JMOngoingConferenceModule.java
@@ -1,8 +1,14 @@
 package org.jitsi.meet.sdk;
 
+import static android.Manifest.permission.POST_NOTIFICATIONS;
+import static android.Manifest.permission.RECORD_AUDIO;
+
 import android.content.Context;
+import android.content.pm.PackageManager;
+import android.os.Build;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,8 +16,13 @@ import com.facebook.react.bridge.ReactApplicationContext;
 import com.facebook.react.bridge.ReactContextBaseJavaModule;
 import com.facebook.react.bridge.ReactMethod;
 import com.facebook.react.module.annotations.ReactModule;
+import com.facebook.react.modules.core.PermissionListener;
 
 import org.jitsi.meet.sdk.log.JitsiMeetLogger;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 
 /**
@@ -23,15 +34,76 @@ class JMOngoingConferenceModule extends ReactContextBaseJavaModule {
 
     public static final String NAME = "JMOngoingConference";
 
+    private static final int PERMISSIONS_REQUEST_CODE = (int) (Math.random() * Short.MAX_VALUE);
+
     public JMOngoingConferenceModule(ReactApplicationContext reactContext) {
         super(reactContext);
     }
 
+    @RequiresApi(api = Build.VERSION_CODES.TIRAMISU)
     @ReactMethod
     public void launch() {
         Context context = getReactApplicationContext();
         ReactActivity reactActivity = (ReactActivity) getCurrentActivity();
-        JMOngoingConferenceService.launch(context, reactActivity);
+
+        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.TIRAMISU) {
+            JMOngoingConferenceService.launch(context);
+            JitsiMeetLogger.i(NAME + " Launching service");
+        }
+
+        PermissionListener listener = new PermissionListener() {
+            @Override
+            public boolean onRequestPermissionsResult(int i, String[] strings, int[] results) {
+                JitsiMeetLogger.i(NAME + " Permission callback received");
+
+                if (results == null || results.length == 0) {
+                    JitsiMeetLogger.w(NAME + " Permission results are null or empty");
+                    return true;
+                }
+
+                int counter = 0;
+                for (int result : results) {
+                    if (result == PackageManager.PERMISSION_GRANTED) {
+                        counter++;
+                    }
+                }
+
+                JitsiMeetLogger.i(NAME + " Permissions granted: " + counter + "/" + results.length);
+
+                if (counter == results.length) {
+                    JitsiMeetLogger.i(NAME + " All permissions granted, launching service");
+                    JMOngoingConferenceService.launch(context);
+                } else {
+                    JitsiMeetLogger.w(NAME + " Not all permissions were granted");
+                }
+
+                return true;
+            }
+        };
+
+        JitsiMeetLogger.i(NAME + " Checking Tiramisu permissions");
+
+        List<String> permissionsList = new ArrayList<>();
+
+        permissionsList.add(POST_NOTIFICATIONS);
+        permissionsList.add(RECORD_AUDIO);
+
+        String[] permissionsArray = new String[ permissionsList.size() ];
+        permissionsArray = permissionsList.toArray( permissionsArray );
+
+        if (permissionsArray.length > 0) {
+            try {
+                JitsiMeetLogger.i(NAME + " Requesting permissions: " + Arrays.toString(permissionsArray));
+                if (reactActivity != null) {
+                    reactActivity.requestPermissions(permissionsArray, PERMISSIONS_REQUEST_CODE, listener);
+                }
+            } catch (Exception e) {
+                JitsiMeetLogger.e(e, NAME + " Error requesting permissions");
+            }
+        } else {
+            JitsiMeetLogger.i(NAME + " No permissions needed, launching service");
+            JMOngoingConferenceService.launch(context);
+        }
 
         JitsiMeetLogger.w(NAME + " launch");
     }
